backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
blinker==1.9.0
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
docopt==0.6.2
Flask==3.1.0
frozenlist==1.5.0
gotrue==2.11.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
opencv-python==4.11.0.86
packaging==24.2
pipreqs==0.4.13
postgrest==0.19.1
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.1.0
requests==2.32.3
s3transfer==0.11.1
six==1.17.0
sniffio==1.3.1
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
Werkzeug==3.1.3
yarg==0.1.10
yarl==1.18.3
